Lots of progress related to monster pathfinding
Specifically, divide the coordinate space into classic pac-man
"frames" which can be used to help with ghost targeting algorithms. At
worst, at this moment we have the ability to tell a given ghost to
strike a random frame within some arbitrary distance of pac-man. The
next step though will be to mark specific frames as "traversable" or
not. It would be nice if this could be done by leveraging the
pathfinding API rather than having to figure it out from scratch.