Sub out high res ghost textures
Also, get behaviors mapped for each ghost. With that said, having read
the article here:
http://gameinternals.com/post/2072558330/understanding-pac-man-ghost-behavior,
it might actually be _simpler_ to implement ghost targetting/movement
behavior on a frame-by-frame basis rather than relying on pathfinding.
The idea would be to use pathfinding/navmeshes to allow the ghosts to
walk properly and not move through walls, but their pathfinding
destination at any given frame will always be 1 frame in front of the
ghost.