Baseline behavior working!!
Turns out that my original intuition was correct. By asking the agent to
update it's target every tick, it was re-evaluating _distances_ every
tick, rather than waiting until it had reached the middle of the next
frame, which is the _correct_ place to make that calculation. The
solution was to simply wait for the ghost to get to the end of their nav
route before selecting a new target.